requirements.txt
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
arrow==1.3.0
binaryornot==0.4.4
boto3==1.28.62
botocore==1.31.62
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.1
click==8.1.7
cookiecutter==2.4.0
python-dotenv==1.0.0
exceptiongroup==1.1.3
fastapi==0.103.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
psycopg2==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0.1
requests==2.31.0
rich==13.6.0
s3transfer==0.7.0
sniffio==1.3.0
SQLAlchemy==2.0.22
sqlalchemy-filters==0.13.0
starlette==0.27.0
text-unidecode==1.3
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
urllib3==1.26.17
uvicorn==0.23.2
